requirements.txt
absl-py==2.3.1
aiocfscrape==1.0.0
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.4.0
aiohttp==3.11.14
aiohttp-proxy==0.1.2
aiohttp_socks==0.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
assemblyai==0.43.1
asyncio==3.4.3
attrs==24.2.0
axios==0.4.0
beautifulsoup4==4.12.3
better-proxy==1.2.0
bitarray==2.9.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
ccxt==4.3.48
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
colorlog==6.8.2
comtypes==1.4.5
contourpy==1.3.3
crayons==0.4.0
cryptography==36.0.1
cssselect==1.2.0
curl_cffi==0.10.0
cycler==0.12.1
cytoolz==1.0.0
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
eth-account==0.13.5
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.0.1
fake-useragent==1.5.1
Faker==27.0.0
findspark==2.0.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.4.1
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
httpx-socks==0.9.1
Hydrogram==0.2.0
idna==2.10
importlib_metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.4
Js2Py==0.70
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyboard==0.13.5
kiwisolver==1.4.9
loguru==0.7.2
lru-dict==1.3.0
lxml==5.3.0
lxml_html_clean==0.2.2
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.1.0
Naked==0.1.32
node==1.2.2
numpy==1.26.4
odict==1.9.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parse==1.20.2
parsimonious==0.9.0
pillow==10.4.0
playsound==1.2.2
plumber==1.7
primp==0.14.0
propcache==0.3.1
protobuf==4.25.8
psutil==6.1.0
py4j==0.10.9.9
pyaes==1.6.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycares==4.6.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.5.2
pydantic_core==2.27.2
pyee==11.1.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyhttpx==2.10.9
pyjsparser==2.7.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pynocaptcha==2.0.45
pynput==1.7.7
pyOpenSSL==21.0.0
pyparsing==3.1.4
pyperclip==1.9.0
pypiwin32==223
pyppeteer==2.0.0
pyquery==2.0.1
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-socks==2.5.2
pyttsx3==2.99
pytweening==1.2.0
pytz==2024.1
pyunormalize==16.0.0
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.0.1
rpds-py==0.20.0
rsa==4.8
scipy==1.16.1
seaborn==0.13.2
selenium==4.25.0
sentencepiece==0.2.1
setproctitle==1.3.3
setuptools==75.1.0
shellescape==3.8.1
six==1.16.0
smart-airdrop-claimer==0.1.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.6
source==1.2.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tabulate==0.9.0
Telethon==1.36.0
termcolor==2.4.0
toolz==1.0.0
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.2.0
urllib3==2.3.0
w3lib==2.2.1
web3==7.8.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.19.0
zipp==3.20.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==6.0
